Fully functioning employee maintenance
Added refresh button too

+++ employee-maintenance.php
@@ -24,14 +24,11 @@
                     <div class="row m-0 gap-3">  
                         <div class="grey_container col col-12-sm p-0 mx-0 text-center justify-content-evenly">
                             <div class="col col-12 card p-3">
-                                <button type="button" class="register_button btn me-3 btn-outline-primary"><a class="nav-link" href="employee-registration.php">Register Employee</a></button>
-
-                                <?php 
-                                    // include('employee-search.php');
-                                ?>
-                                <form class="d-flex" method="get" action="employee-search.php">
+                                <form class="d-flex" method="get">
+                                    <button type="button" class="register_button btn me-3 btn-outline-primary"><a class="nav-link" href="employee-registration.php">Register Employee</a></button>
                                     <input class="form-control me-3" type="text" placeholder="Enter employee name or id" name="emp_id">
                                     <input class="search_button btn btn-outline-success" type="submit" value="Search">
+                                    <a href="employee-maintenance.php"><button type="button" class="btn"><img id="exit-button-img" src="images/refresh.png"></button></a>
                                 </form>
 
 
@@ -57,34 +54,72 @@
                                 <?php
                                     include "connection.php";
                                     
-                                    $sql = "SELECT * FROM employee";
-                                    $result = mysqli_query($conn, $sql);
-                                    
-                                    if (mysqli_num_rows($result) > 0) {
-                                        while ($row = mysqli_fetch_assoc($result)) {
-                                            echo "<tr>";
-                                            echo "<td>" . $row["emp_id"] . "</td>";
-                                            echo "<td>" . $row["full_name"] . "</td>";
-                                            echo "<td>" . $row["contact_number"] . "</td>";
-                                            echo "<td>" . $row["email_address"] . "</td>";
-                                            echo "<td>" . $row["address"] . "</td>";
-                                            echo "<td>" . $row["zip"] . "</td>";
-                                            echo "<td>" . $row["contract"] . "</td>";
-                                            echo "<td>" . $row["shift"] . "</td>";
-                                            echo "<td>
-                                                    <div class='btn_container me-0'>
-                                                        <button class='btn btn-outline-dark'><a href='employee-edit.php?id=" . $row["emp_id"] . "' class='nav-link'>Edit</a></button>
-                                                        <button class='btn btn-outline-danger'><a href='delete_employee.php?id=" . $row["emp_id"] . "' class='nav-link'>Delete</a></button>
-                                                    </div>
-                                                </td>";
-                                            echo "</tr>";
+                                    if (isset($_GET["emp_id"])) {
+                                        $emp_id = $_GET["emp_id"];
+                                
+                                        if (!$emp_id) {
+                                            echo "<script> ('Employeee ID Not Specificied!');  window.location.href = 'employee-maintenance.php'</script>";
+                                            return;
+                                        }
+                                
+                                        $sql = "SELECT * FROM employee WHERE emp_id = '" . $emp_id . "' ";
+                                        $result = mysqli_query($conn, $sql);
+                                
+                                        if (!$result) {
+                                            echo "<p class='error'>Error: " . mysqli_error() . "</p>";
+                                            return;
+                                        }
+                                
+                                        if (mysqli_num_rows($result) > 0) {
+                                            while ($row = mysqli_fetch_assoc($result)) {
+                                                echo "<tr>";
+                                                echo "<td>" . $row["emp_id"] . "</td>";
+                                                echo "<td>" . $row["full_name"] . "</td>";
+                                                echo "<td>" . $row["contact_number"] . "</td>";
+                                                echo "<td>" . $row["email_address"] . "</td>";
+                                                echo "<td>" . $row["address"] . "</td>";
+                                                echo "<td>" . $row["zip"] . "</td>";
+                                                echo "<td>" . $row["contract"] . "</td>";
+                                                echo "<td>" . $row["shift"] . "</td>";
+                                                echo "<td>
+                                                        <div class='btn_container me-0'>
+                                                            <button class='btn btn-outline-dark'><a href='employee-edit.php?id=" . $row["emp_id"] . "' class='nav-link'>Edit</a></button>
+                                                            <button class='btn btn-outline-danger'><a href='delete_employee.php?id=" . $row["emp_id"] . "' class='nav-link'>Delete</a></button>
+                                                        </div>
+                                                    </td>";
+                                                echo "</tr>";
+                                            }
+                                        } else {
+                                            echo "<tr><td colspan='9'>Employee not found!</td></tr>";
                                         }
                                     } else {
-                                        echo "No employees found.";
-                                    }
-
+                                        $sql = "SELECT * FROM employee";
+                                        $result = mysqli_query($conn, $sql);
+                                
+                                        if (mysqli_num_rows($result) > 0) {
+                                            while ($row = mysqli_fetch_assoc($result)) {
+                                                echo "<tr>";
+                                                echo "<td>" . $row["emp_id"] . "</td>";
+                                                echo "<td>" . $row["full_name"] . "</td>";
+                                                echo "<td>" . $row["contact_number"] . "</td>";
+                                                echo "<td>" . $row["email_address"] . "</td>";
+                                                echo "<td>" . $row["address"] . "</td>";
+                                                echo "<td>" . $row["zip"] . "</td>";
+                                                echo "<td>" . $row["contract"] . "</td>";
+                                                echo "<td>" . $row["shift"] . "</td>";
+                                                echo "<td>
+                                                        <div class='btn_container me-0'>
+                                                            <button class='btn btn-outline-dark'><a href='employee-edit.php?id=" . $row["emp_id"] . "' class='nav-link'>Edit</a></button>
+                                                            <button class='btn btn-outline-danger'><a href='delete_employee.php?id=" . $row["emp_id"] . "' class='nav-link'>Delete</a></button>
+                                                        </div>
+                                                    </td>";
+                                                echo "</tr>";
+                                            }
+                                        } else {
+                                            echo "No employees found.";
+                                        }
+                                    } 
                                     $conn->close();
-
                                 ?>
                             </tbody>
                             </table>
